+++ MangaMark_Vue/src/factory/websiteFactory.ts
@@ -43,6 +43,7 @@ export class WebsiteFactory {
         return new AqumangaWebsite()
       case domain.includes('toptruyen'):
       case domain.includes('doctruyen3q'):
+      case domain.includes('truyengg'):
         return new ToptruyenWebsite()
       case domain.includes('manganato'):
         return new ManganatoWebsite()