+++ MangaMark_Vue/src/js/website/cmangaWebsite.ts
@@ -61,6 +61,9 @@ export default class CmangaWebsite implements Website {
   blockAds(): void {
     removeElements('.pr_module');
     removeElements('img[src*="cmangapi"]')
+    removeElements('div#popup_content')
+    removeElements('div.chapter_ad_block')
+    removeElements('div.chapter_content', 'style')
     removeCookie('ads_num')
     removeCookie('adsView')
   }